refactor: move audio package from utils to top
