+++ setup.py
@@ -51,13 +51,13 @@ setup(
     license='MIT License',
     packages=[
         'spleeter',
+        'spleeter.audio',
         'spleeter.commands',
         'spleeter.model',
         'spleeter.model.functions',
         'spleeter.model.provider',
         'spleeter.resources',
         'spleeter.utils',
-        'spleeter.utils.audio',
     ],
     package_data={'spleeter.resources': ['*.json']},
     python_requires='>=3.6, <3.8',
