+++ spleeter/commands/train.py
@@ -13,9 +13,9 @@ from functools import partial
 import tensorflow as tf
 # pylint: enable=import-error
 
+from ..audio.adapter import get_audio_adapter
 from ..dataset import get_training_dataset, get_validation_dataset
 from ..model import model_fn
-from ..utils.audio.adapter import get_audio_adapter
 from ..utils.logging import get_logger
 
 __email__ = 'research@deezer.com'
